requirements.txt
numpy==2.3.5
torch==2.9.1
transformers==4.57.1
pandas==2.3.3
scikit-learn==1.7.2
tqdm==4.67.1
psutil==7.1.3
py-cpuinfo==9.0.0
matplotlib==3.10.7
scipy==1.16.3
seaborn==0.13.2
textual==6.7.1
rich==14.2.0